Django/requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
django-cors-headers==3.13.0
django_plotly_dash==2.0.0
psycopg2==2.9.3
channels==2.4.0
bootstrap4==0.1.0
dpd_static_support==0.0.5
dash_renderer==1.9.1
plotly==5.10.0
whitenoise==6.2.0
pandas==1.5.0